+++ graphics/stage/main/area_roulette_selected_fx.png
