+++ include/functions.h
@@ -534,16 +534,16 @@ extern void DrawSpoinkSprite();
 extern void RunEvolutionCutscene(void);
 //extern ? sub_2DE54();
 extern void RunTravelEventCutscene(void);
-extern void DecrementFieldTimer();
+extern void DecrementPelipperTimer();
 extern void UpdatePelipperPondEntity();
 extern void AnimateWailmerEntity();
 extern void UpdateZigzagoonEntity();
-extern void DrawZigzagoonAndShockWall();
+extern void DrawZigzagoonAndRouletteStopPrompt();
 extern void UpdateSapphireBumperLogic();
 extern void DrawSapphireBumperSprites();
 extern void CalculateRubyBumperBounce();
 extern void HandleRubyBumperHit();
-extern void InitSapphireEggCaveState();
+extern void InitSapphireEggHatchState();
 extern void UpdateSapphireEggHatchAnimation();
 extern void UpdateSapphireSeedotCollection();
 extern void DrawSapphireSeedotAndBasketSprites();
