+++ src/rom_27E08.c
@@ -2,40 +2,7 @@
 #include "m4a.h"
 #include "main.h"
 #include "constants/bg_music.h"
-#include "constants/ruby_states.h"
 
-#define BONUS_CATCH_TIME 7200 //2 minutes, 60FPS
-
-extern const u8 gDefaultBallPalette[];
-extern const u8 *gEvoItemAppear_GfxList[];
-extern const u8 gEvoItemPalettes[][0x20];
-extern const s16 gEvoItemAppearFrameThresholds[];
-extern const u16 gEvoItemAnimOamFramesets[58][15];
-extern const struct Vector16 gEvoItemPositions[][8];
-extern const s16 gBumperAnimFrames[];
-extern const s16 gPondDialAnimFrames[];
-extern const s16 gAreaRouletteTable[][7];
-extern const s16 gAreaToSpeciesTable[];
-extern const u16 gAreaRouletteOamFramesets[18][27];
-extern const u8 gHatchCompletePalette[];
-extern const u8 gEggPortraitTilesGfx[];
-extern const u8 gHatchStartTilesGfx[];
-extern const u8 gHatchStartPalette[];
-extern const s16 gHatchRevealSparkleTimings[];
-extern const u16 gHatchRevealOamFramesets[16][18];
-extern const u8 gHatchStage2TilesGfx[];
-extern const u8 gHatchStage2Palette[];
-extern const u16 gHatchFullRevealOamFramesets[18][18];
-extern const u8 gHatchStage3TilesGfx[];
-extern const u8 gHatchStage3Palette[];
-extern const u8 gHatchFinalTilesGfx[];
-extern const u8 gHatchFinalPalette[];
-extern const s16 gHatchRevealFinalTimings[];
-extern const u16 gHatchSequentialOamFramesets[22][12];
-extern const u8 gHatchRevealPalette[];
-extern const u8 gHatchRevealTilesGfx[];
-extern const s16 gHatchSequentialTileFramesetData[][2];
-extern const u16 gHatchSequentialTileBreakSpritesheetOam[28][18];
 extern const s16 gEggHatchAnimData[28][3];
 extern const u8 gCatchSpriteFrameBuffer[][0x120];
 extern const struct Vector32 gSapphireEggWaypoints[];
@@ -75,9 +42,7 @@ extern const u16 gAngleToDirectionTable[];
 
 extern u8 gCatchSpritePaletteBuffer[];
 
-extern struct SongHeader se_evo_item_appear;
-extern struct SongHeader se_evo_item_finish_appear;
-extern struct SongHeader se_evo_item_collected;
+
 
 void ResetCatchState(s16 resetHoleIndicators)
 {
@@ -2040,1408 +2005,5 @@ void UpdateEggMode(void)
             gCurrentPinballGame->boardSubState = 0;
         }
         break;
-    }
-}
-
-void DisableHatchTileDisplay(void)
-{
-    s16 i;
-    struct SpriteGroup *group;
-    struct OamDataSimple *oamSimple;
-
-    group = gMain.fieldSpriteGroups[18];
-    if (group->available)
-    {
-        for (i = 0; i < 6; i++)
-        {
-            oamSimple = &group->oam[i];
-            gOamBuffer[oamSimple->oamId].x = 200;
-            gOamBuffer[oamSimple->oamId].y = 180;
-        }
-    }
-    gMain.fieldSpriteGroups[18]->available = 0;
-    group = gMain.fieldSpriteGroups[12];
-    if (group->available)
-    {
-        for (i = 0; i < 6; i++)
-        {
-            oamSimple = &group->oam[i];
-            gOamBuffer[oamSimple->oamId].x = 200;
-            gOamBuffer[oamSimple->oamId].y = 180;
-        }
-    }
-    gMain.fieldSpriteGroups[12]->available = 0;
-    gCurrentPinballGame->activePortraitType = 0;
-}
-
-void RevealSequentialHatchTiles(void)
-{
-    s16 i;
-    struct SpriteGroup *group;
-    struct OamDataSimple *oamSimple;
-    u16 *dst;
-    s16 var0;
-    int var1;
-
-    var1 = 1;
-    gMain.fieldSpriteGroups[18]->available = 1;
-    if (gCurrentPinballGame->hatchFrameId > 0)
-        UpdateSequentialTileParticles();
-
-    if (gCurrentPinballGame->hatchSequentialTilesRevealed < gCurrentPinballGame->hatchTilesBoardAcknowledged)
-    {
-        if (gCurrentPinballGame->hatchFrameId == 0 && gCurrentPinballGame->hatchSequentialTileRevealFrameAnimTimer == 0)
-        {
-            gCurrentPinballGame->activePortraitType = 11;
-            DmaCopy16(3, gHatchRevealPalette, (void *)0x050003C0, 0x20);
-            DmaCopy16(3, gHatchRevealTilesGfx, (void *)0x06015800, 0x2800);
-        }
-
-        if (gHatchSequentialTileFramesetData[gCurrentPinballGame->hatchFrameId][1] > gCurrentPinballGame->hatchSequentialTileRevealFrameAnimTimer)
-        {
-            gCurrentPinballGame->hatchSequentialTileRevealFrameAnimTimer++;
-        }
-        else
-        {
-            gCurrentPinballGame->hatchSequentialTileRevealFrameAnimTimer = 0;
-            gCurrentPinballGame->hatchFrameId++;
-            if (gCurrentPinballGame->hatchFrameId > 12)
-            {
-                gCurrentPinballGame->hatchSequentialTilesRevealed++;
-                gCurrentPinballGame->hatchFrameId = 0;
-                var1 = 0;
-                gCurrentPinballGame->hatchTilesRemaining--;
-                if (gCurrentPinballGame->hatchSequentialTilesRevealed == gCurrentPinballGame->hatchTilesBoardAcknowledged)
-                {
-                    if (gCurrentPinballGame->hatchTilesRemaining >= 0)
-                        gCurrentPinballGame->activePortraitType = 0;
-                }
-
-                if (gCurrentPinballGame->hatchTilesRemaining < 0)
-                {
-                    gCurrentPinballGame->hatchTilesBoardAcknowledged = 0;
-                    gCurrentPinballGame->hatchSequentialTilesRevealed = 0;
-                    gCurrentPinballGame->hatchTilesBumperAcknowledged = 0;
-                    for (i = 0; i < 6; i++)
-                        gCurrentPinballGame->hatchTileShufflePool[i] = i;
-
-                    var0 = gMain.systemFrameCount % 6;
-                    gCurrentPinballGame->hatchTilesRemaining = 5;
-                    gCurrentPinballGame->boardSubState++;
-                    gMain.fieldSpriteGroups[18]->available = 0;
-                }
-                else if (gCurrentPinballGame->hatchTilesRemaining == 0)
-                {
-                    var0 = 0;
-                }
-                else
-                {
-                    var0 = gMain.systemFrameCount % gCurrentPinballGame->hatchTilesRemaining;
-                }
-
-                gCurrentPinballGame->hatchGridCellIndex = gCurrentPinballGame->hatchTileShufflePool[var0];
-                for (i = var0; i < gCurrentPinballGame->hatchTilesRemaining; i++)
-                    gCurrentPinballGame->hatchTileShufflePool[i] = gCurrentPinballGame->hatchTileShufflePool[i + 1];
-            }
-            else
-            {
-                if (gCurrentPinballGame->hatchFrameId == 1)
-                    gCurrentPinballGame->hatchTilePalette[gCurrentPinballGame->hatchGridCellIndex] = 13;
-            }
-        }
-    }
-    else
-    {
-        var1 = 0;
-    }
-
-    if (gCurrentPinballGame->hatchFrameId == 1)
-    {
-        m4aSongNumStart(SE_CATCH_TILE_REVEAL);
-        InitSequentialTileParticles();
-    }
-
-    if (gCurrentPinballGame->hatchFrameId == 12)
-        var1 = 0;
-
-    group = gMain.fieldSpriteGroups[18];
-    if (var1)
-    {
-        group->baseX = (gCurrentPinballGame->hatchGridCellIndex % 3) * 16 - (gCurrentPinballGame->cameraXOffset - 96);
-        group->baseY = (gCurrentPinballGame->hatchGridCellIndex / 3) * 16 - (gCurrentPinballGame->cameraYOffset - 300);
-    }
-    else
-    {
-        group->baseY = 200;
-    }
-
-    if (group->baseY >= 200)
-        group->baseY = 200;
-
-    for (i = 0; i < 6; i++)
-    {
-        oamSimple = &group->oam[i];
-        dst = (u16 *)&gOamBuffer[oamSimple->oamId];
-        *dst++ = gHatchSequentialTileBreakSpritesheetOam[gHatchSequentialTileFramesetData[gCurrentPinballGame->hatchFrameId][0]][i * 3 + 0];
-        *dst++ = gHatchSequentialTileBreakSpritesheetOam[gHatchSequentialTileFramesetData[gCurrentPinballGame->hatchFrameId][0]][i * 3 + 1];
-        *dst++ = gHatchSequentialTileBreakSpritesheetOam[gHatchSequentialTileFramesetData[gCurrentPinballGame->hatchFrameId][0]][i * 3 + 2];
-
-        gOamBuffer[oamSimple->oamId].priority = 3;
-        gOamBuffer[oamSimple->oamId].x += group->baseX;
-        gOamBuffer[oamSimple->oamId].y += group->baseY;
-    }
-}
-
-void RevealAllHatchTilesAtOnce(void)
-{
-    s16 i;
-    struct SpriteGroup *group;
-    struct OamDataSimple *oamSimple;
-    u16 *dst;
-    s16 var0;
-
-    switch (gCurrentPinballGame->hatchRevealPhase)
-    {
-    case 0:
-        if (gCurrentPinballGame->revealAnimFrameCounter < 60)
-        {
-            gCurrentPinballGame->revealAnimFrameCounter++;
-        }
-        else
-        {
-            gCurrentPinballGame->hatchRevealPhase++;
-            gCurrentPinballGame->revealAnimFrameCounter = 0;
-        }
-        gMain.blendControl = 0xCE;
-        break;
-    case 1:
-        gCurrentPinballGame->activePortraitType = 5;
-        DmaCopy16(3, gHatchStartTilesGfx, (void *)0x06015800, 0x2000);
-        DmaCopy16(3, gHatchStartPalette, (void *)0x050003C0, 0x20);
-        gMain.fieldSpriteGroups[35]->available = 1;
-        m4aSongNumStart(SE_CATCH_ALL_REVEAL_LIGHTNING);
-        gCurrentPinballGame->hatchRevealPhase++;
-        gCurrentPinballGame->revealAnimFrameCounter = 0;
-        gCurrentPinballGame->revealFramesetIndex = 0;
-        break;
-    case 2:
-        if (gHatchRevealSparkleTimings[gCurrentPinballGame->revealFramesetIndex] > gCurrentPinballGame->revealAnimFrameCounter)
-        {
-            gCurrentPinballGame->revealAnimFrameCounter++;
-        }
-        else
-        {
-            gCurrentPinballGame->revealAnimFrameCounter = 0;
-            gCurrentPinballGame->revealFramesetIndex++;
-            if (gCurrentPinballGame->revealFramesetIndex > 7)
-            {
-                gCurrentPinballGame->revealFramesetIndex = 7;
-                gCurrentPinballGame->hatchRevealPhase++;
-                gMain.fieldSpriteGroups[35]->available = 0;
-            }
-        }
-
-        var0 = gCurrentPinballGame->revealFramesetIndex;
-        group = gMain.fieldSpriteGroups[35];
-        group->baseX = 124 - gCurrentPinballGame->cameraXOffset;
-        group->baseY = 244 - gCurrentPinballGame->cameraYOffset;
-        if (group->baseY >= 200)
-            group->baseY = 200;
-
-        for (i = 0; i < 6; i++)
-        {
-            oamSimple = &group->oam[i];
-            dst = (u16 *)&gOamBuffer[oamSimple->oamId];
-            *dst++ = gHatchRevealOamFramesets[var0][i * 3 + 0];
-            *dst++ = gHatchRevealOamFramesets[var0][i * 3 + 1];
-            *dst++ = gHatchRevealOamFramesets[var0][i * 3 + 2];
-
-            gOamBuffer[oamSimple->oamId].x += group->baseX;
-            gOamBuffer[oamSimple->oamId].y += group->baseY;
-        }
-        break;
-    case 3:
-        gCurrentPinballGame->activePortraitType = 6;
-        DmaCopy16(3, gHatchStage2TilesGfx, (void *)0x06015800, 0x800);
-        DmaCopy16(3, gHatchStage2Palette, (void *)0x050003C0, 0x20);
-        gMain.fieldSpriteGroups[36]->available = 1;
-        gCurrentPinballGame->hatchRevealPhase++;
-        gCurrentPinballGame->revealAnimFrameCounter = 0;
-        gCurrentPinballGame->startButtonDisabled = 1;
-        break;
-    case 4:
-        var0 = gCurrentPinballGame->revealAnimFrameCounter / 2;
-        if (gCurrentPinballGame->revealAnimFrameCounter < 15)
-        {
-            gCurrentPinballGame->revealAnimFrameCounter++;
-        }
-        else
-        {
-            gMain.fieldSpriteGroups[36]->available = 0;
-            gCurrentPinballGame->hatchRevealPhase++;
-            gCurrentPinballGame->revealAnimFrameCounter = 0;
-            gCurrentPinballGame->revealFramesetIndex = 0;
-        }
-
-        group = gMain.fieldSpriteGroups[36];
-        group->baseX = 96 - gCurrentPinballGame->cameraXOffset;
-        group->baseY = 300 - gCurrentPinballGame->cameraYOffset;
-        if (group->baseY >= 200)
-            group->baseY = 200;
-
-        for (i = 0; i < 6; i++)
-        {
-            oamSimple = &group->oam[i];
-            dst = (u16 *)&gOamBuffer[oamSimple->oamId];
-            *dst++ = gHatchFullRevealOamFramesets[var0][i * 3 + 0];
-            *dst++ = gHatchFullRevealOamFramesets[var0][i * 3 + 1];
-            *dst++ = gHatchFullRevealOamFramesets[var0][i * 3 + 2];
-
-            gOamBuffer[oamSimple->oamId].x += group->baseX;
-            gOamBuffer[oamSimple->oamId].y += group->baseY;
-        }
-        break;
-    case 5:
-        gCurrentPinballGame->activePortraitType = 7;
-        DmaCopy16(3, gHatchStage3TilesGfx, (void *)0x06015800, 0x2000);
-        DmaCopy16(3, gHatchStage3Palette, (void *)0x050003C0, 0x20);
-        gCurrentPinballGame->hatchRevealPhase++;
-        InitBurstTileParticles();
-        m4aSongNumStart(SE_CATCH_ALL_REVEAL_SHATTER);
-        break;
-    case 6:
-        gMain.blendBrightness = 0;
-        gMain.blendControl = ((REG_DISPCNT & (DISPCNT_OBJ_ON | DISPCNT_BG_ALL_ON)) >> 8) | 0xA0;
-        if (gCurrentPinballGame->revealAnimFrameCounter < 73)
-        {
-            s16 var1;
-            if (gCurrentPinballGame->revealAnimFrameCounter < 8)
-            {
-                gMain.blendBrightness = 16;
-            }
-            else
-            {
-                for (i = 0; i < 6; i++)
-                    gCurrentPinballGame->hatchTilePalette[i] = 13;
-
-                gMain.blendBrightness = ((8 - gCurrentPinballGame->revealAnimFrameCounter) / 4) + 16;
-            }
-        }
-
-        if (gCurrentPinballGame->revealAnimFrameCounter < 72)
-        {
-            gCurrentPinballGame->revealAnimFrameCounter++;
-        }
-        else
-        {
-            gCurrentPinballGame->hatchRevealPhase++;
-            gCurrentPinballGame->revealAnimFrameCounter = 0;
-            gCurrentPinballGame->revealFramesetIndex = 0;
-        }
-
-        if (gCurrentPinballGame->revealAnimFrameCounter > 8U)
-        {
-            if (gCurrentPinballGame->hatchSequentialTileRevealFrameAnimTimer < 0x7000)
-                UpdateBurstTileParticles();
-        }
-        break;
-    case 7:
-        gCurrentPinballGame->activePortraitType = 8;
-        DmaCopy16(3, gHatchFinalTilesGfx, (void *)0x06015800, 0x1800);
-        DmaCopy16(3, gHatchFinalPalette, (void *)0x050003C0, 0x20);
-        gMain.fieldSpriteGroups[37]->available = 1;
-        gCurrentPinballGame->hatchRevealPhase++;
-        gCurrentPinballGame->startButtonDisabled = 0;
-        break;
-    case 8:
-        if (gHatchRevealFinalTimings[gCurrentPinballGame->revealFramesetIndex] > gCurrentPinballGame->revealAnimFrameCounter)
-        {
-            gCurrentPinballGame->revealAnimFrameCounter++;
-        }
-        else
-        {
-            gCurrentPinballGame->revealAnimFrameCounter = 0;
-            gCurrentPinballGame->revealFramesetIndex++;
-            if (gCurrentPinballGame->revealFramesetIndex > 10)
-            {
-                gCurrentPinballGame->boardSubState++;
-                gMain.fieldSpriteGroups[37]->available = 0;
-                gCurrentPinballGame->hatchTilesBoardAcknowledged = 0;
-                gCurrentPinballGame->hatchSequentialTilesRevealed = 0;
-                gCurrentPinballGame->hatchTilesBumperAcknowledged = 0;
-                gMain.blendControl = 0xCE;
-                gMain.blendBrightness = 0;
-                gMain.fieldSpriteGroups[37]->available = 0;
-                gCurrentPinballGame->revealFramesetIndex = 10;
-                gCurrentPinballGame->activePortraitType = 0;
-            }
-        }
-
-        var0 = gCurrentPinballGame->revealFramesetIndex;
-        group = gMain.fieldSpriteGroups[37];
-        group->baseX = 96 - gCurrentPinballGame->cameraXOffset;
-        group->baseY = 300 - gCurrentPinballGame->cameraYOffset;
-        if (group->baseY >= 200)
-            group->baseY = 200;
-
-        for (i = 0; i < 4; i++)
-        {
-            oamSimple = &group->oam[i];
-            dst = (u16 *)&gOamBuffer[oamSimple->oamId];
-            *dst++ = gHatchSequentialOamFramesets[var0][i * 3 + 0];
-            *dst++ = gHatchSequentialOamFramesets[var0][i * 3 + 1];
-            *dst++ = gHatchSequentialOamFramesets[var0][i * 3 + 2];
-
-            gOamBuffer[oamSimple->oamId].x += group->baseX;
-            gOamBuffer[oamSimple->oamId].y += group->baseY;
-        }
-        break;
-    }
-}
-
-//start 'roulette.c' here
-void InitAreaRoulette(void)
-{
-    s16 i;
-
-    gCurrentPinballGame->boardSubState = 1;
-    gCurrentPinballGame->stageTimer = 0;
-    gCurrentPinballGame->creatureOamPriority = 3;
-    gCurrentPinballGame->areaVisitCount = 0;
-    gCurrentPinballGame->areaRouletteSlotIndex = (Random() + gMain.systemFrameCount) % 6;
-    gCurrentPinballGame->area = gAreaRouletteTable[gMain.selectedField][gCurrentPinballGame->areaRouletteSlotIndex];
-    gCurrentPinballGame->rouletteAreaIndex[1] = gAreaToSpeciesTable[gCurrentPinballGame->area];
-    gCurrentPinballGame->area = gAreaRouletteTable[gMain.selectedField][(gCurrentPinballGame->areaRouletteSlotIndex + 1) % 6];
-    gCurrentPinballGame->rouletteAreaIndex[0] = gAreaToSpeciesTable[gCurrentPinballGame->area];
-    LoadPortraitGraphics(0, 0);
-    LoadPortraitGraphics(0, 1);
-    for (i = 0; i < 6; i++)
-        gCurrentPinballGame->hatchTilePalette[i] = 13;
-
-    DmaCopy16(3, gHatchCompletePalette, (void *)0x050003C0, 0x20);
-    DmaCopy16(3, gEggPortraitTilesGfx, (void *)0x06015800, 0x280);
-    gCurrentPinballGame->activePortraitType = 12;
-}
-
-void UpdateAreaRoulette(void)
-{
-    s16 i;
-    struct SpriteGroup *group;
-    struct OamDataSimple *oamSimple;
-    u16 *dst;
-    s16 var0;
-
-    switch (gCurrentPinballGame->boardSubState)
-    {
-    case 1:
-        gMain.blendControl = 0x1C10;
-        gMain.blendAlpha = BLDALPHA_BLEND(0, 16);
-        gCurrentPinballGame->boardSubState++;
-        gCurrentPinballGame->rouletteSubOffset = 0;
-        gCurrentPinballGame->cameraScrollTarget = 0;
-        gCurrentPinballGame->cameraScrollEnabled = 1;
-        gMain.fieldSpriteGroups[20]->available = 1;
-        gMain.fieldSpriteGroups[21]->available = 1;
-        gMain.fieldSpriteGroups[23]->available = 1;
-        gMain.fieldSpriteGroups[22]->available = 1;
-        gMain.fieldSpriteGroups[19]->available = 1;
-        gCurrentPinballGame->rouletteFrameIndex = 30;
-        gCurrentPinballGame->rouletteRotationPeriod = 30;
-        gCurrentPinballGame->rouletteSpinSpeed = 0;
-        if (gMain.selectedField == FIELD_RUBY)
-        {
-            gCurrentPinballGame->pondBumperStates[0] = 10;
-            gCurrentPinballGame->pondBumperStates[1] = 10;
-            gCurrentPinballGame->pondBumperStates[2] = 10;
-        }
-        else
-        {
-            gCurrentPinballGame->pondBumperStates[0] = 0;
-            gCurrentPinballGame->pondBumperStates[1] = 0;
-            gCurrentPinballGame->pondBumperStates[2] = 0;
-        }
-        gCurrentPinballGame->rubyPondChangeTimer = 0;
-        break;
-    case 2:
-        if (gCurrentPinballGame->cameraScrollOffset == 0)
-        {
-            gCurrentPinballGame->boardSubState++;
-            gCurrentPinballGame->stageTimer = 0;
-        }
-
-        if (gMain.selectedField == FIELD_RUBY)
-        {
-            if (gCurrentPinballGame->rubyPondChangeTimer < 143)
-                gCurrentPinballGame->pondBumperStates[0] = gPondDialAnimFrames[gCurrentPinballGame->rubyPondChangeTimer / 8];
-
-            if (gCurrentPinballGame->rubyPondChangeTimer >= 18 && gCurrentPinballGame->rubyPondChangeTimer < 161)
-                gCurrentPinballGame->pondBumperStates[2] = gPondDialAnimFrames[(gCurrentPinballGame->rubyPondChangeTimer - 18) / 8];
-
-            if (gCurrentPinballGame->rubyPondChangeTimer >= 36 && gCurrentPinballGame->rubyPondChangeTimer < 179)
-                gCurrentPinballGame->pondBumperStates[1] = gPondDialAnimFrames[(gCurrentPinballGame->rubyPondChangeTimer - 36) / 8];
-
-            gCurrentPinballGame->rubyPondChangeTimer++;
-        }
-
-        gCurrentPinballGame->rouletteSubOffset = (gCurrentPinballGame->rouletteFrameIndex * 32) / gCurrentPinballGame->rouletteRotationPeriod;
-        gCurrentPinballGame->portraitDisplayState = 1;
-        gCurrentPinballGame->stageTimer++;
-        UpdateRouletteAnimState();
-        break;
-    case 3:
-        if (gCurrentPinballGame->stageTimer < 15)
-        {
-            gCurrentPinballGame->stageTimer++;
-            if (gCurrentPinballGame->stageTimer == 15)
-            {
-                gCurrentPinballGame->boardSubState++;
-                gCurrentPinballGame->rouletteRotationPeriod = 6;
-                gCurrentPinballGame->rouletteSpinSpeed = 0;
-            }
-        }
-        SetRouletteActiveState(0);
-        /* fallthrough */
-    case 4:
-    case 5:
-        if (gCurrentPinballGame->boardSubState == 3)
-        {
-            gCurrentPinballGame->rouletteFrameIndex--;
-            gCurrentPinballGame->rouletteFrameIndex %= gCurrentPinballGame->rouletteRotationPeriod;
-        }
-        else if (gCurrentPinballGame->boardSubState == 4)
-        {
-            gCurrentPinballGame->rouletteFrameIndex++;
-            if (gCurrentPinballGame->newButtonActions[1] && gCurrentPinballGame->boardSubState == 4)
-            {
-                if (gCurrentPinballGame->rouletteSpinSpeed == 0)
-                {
-                    if (gMain.eReaderBonuses[EREADER_RUIN_AREA_CARD])
-                        gCurrentPinballGame->rouletteSpinSpeed = 48;
-                    else
-                        gCurrentPinballGame->rouletteSpinSpeed = 1;
-
-                    gCurrentPinballGame->rouletteInitialSpeed = gCurrentPinballGame->rouletteSpinSpeed;
-                    gCurrentPinballGame->spoinkEntityState = 1;
-                    gCurrentPinballGame->launcherCharging = 1;
-                    gCurrentPinballGame->saverTimeRemaining = 3600;
-                    gCurrentPinballGame->rubyPondState = RUBY_POND_STATE_CHINCHOU_COUNTERCLOCKWISE;
-                }
-            }
-
-            if (gCurrentPinballGame->rouletteSpinSpeed)
-            {
-                gCurrentPinballGame->rouletteSpinSpeed--;
-                if (gMain.eReaderBonuses[EREADER_RUIN_AREA_CARD])
-                {
-                    gCurrentPinballGame->numCompletedBonusStages = 4;
-                    if (gCurrentPinballGame->rouletteSpinSpeed < 24)
-                    {
-                        gCurrentPinballGame->rouletteSpinSpeed = 24;
-                        if (gCurrentPinballGame->areaRouletteSlotIndex == 6)
-                            gCurrentPinballGame->rouletteSpinSpeed = 0;
-                    }
-                }
-
-                if (gCurrentPinballGame->rouletteSpinSpeed == 0)
-                {
-                    gCurrentPinballGame->boardSubState = 5;
-                }
-                else
-                {
-                    if (gCurrentPinballGame->rouletteFrameIndex == gCurrentPinballGame->rouletteRotationPeriod)
-                    {
-                        gCurrentPinballGame->rouletteRotationPeriod = 40 - (gCurrentPinballGame->rouletteSpinSpeed * 30) / gCurrentPinballGame->rouletteInitialSpeed;
-                        gCurrentPinballGame->rouletteFrameIndex = 0;
-                    }
-                }
-            }
-
-            gCurrentPinballGame->rouletteFrameIndex %= gCurrentPinballGame->rouletteRotationPeriod;
-        }
-        else
-        {
-            gCurrentPinballGame->rouletteFrameIndex++;
-            gCurrentPinballGame->rouletteFrameIndex %= gCurrentPinballGame->rouletteRotationPeriod;
-            if (gCurrentPinballGame->rouletteFrameIndex == 0)
-            {
-                gCurrentPinballGame->stageTimer = 0;
-                gCurrentPinballGame->boardSubState++;
-                gMain.fieldSpriteGroups[23]->available = 0;
-                gMain.fieldSpriteGroups[20]->available = 0;
-                gMain.fieldSpriteGroups[21]->available = 0;
-                gMain.fieldSpriteGroups[34]->available = 1;
-                m4aSongNumStart(SE_AREA_ROULETTE_SELECTED);
-            }
-        }
-
-        gCurrentPinballGame->rouletteSubOffset = (gCurrentPinballGame->rouletteFrameIndex * 32) / gCurrentPinballGame->rouletteRotationPeriod;
-        if (gCurrentPinballGame->rouletteFrameIndex == 0)
-        {
-            gCurrentPinballGame->rouletteAreaIndex[0] = gCurrentPinballGame->rouletteAreaIndex[1];
-            LoadPortraitGraphics(0, 0);
-        }
-
-        if (gCurrentPinballGame->rouletteFrameIndex == 1)
-        {
-            if (gMain.eReaderBonuses[EREADER_RUIN_AREA_CARD])
-            {
-                gCurrentPinballGame->areaRouletteSlotIndex = (gCurrentPinballGame->areaRouletteSlotIndex + 1) % 7;
-                gCurrentPinballGame->areaRouletteNextSlot = 0;
-                gCurrentPinballGame->areaRouletteFarSlot = 1;
-            }
-            else
-            {
-                gCurrentPinballGame->areaRouletteSlotIndex = (gCurrentPinballGame->areaRouletteSlotIndex + 1) % 6;
-                gCurrentPinballGame->areaRouletteNextSlot = (gCurrentPinballGame->areaRouletteSlotIndex + 1) % 6;
-                gCurrentPinballGame->areaRouletteFarSlot = (gCurrentPinballGame->areaRouletteSlotIndex + 2) % 6;
-            }
-
-            gCurrentPinballGame->area = gAreaRouletteTable[gMain.selectedField][gCurrentPinballGame->areaRouletteSlotIndex];
-            gCurrentPinballGame->rouletteAreaIndex[1] = gAreaToSpeciesTable[gCurrentPinballGame->area];
-            LoadPortraitGraphics(0, 1);
-            m4aSongNumStart(SE_ROULETTE_TICK);
-        }
-        break;
-    case 6:
-        if (gCurrentPinballGame->stageTimer < 50)
-        {
-            gCurrentPinballGame->stageTimer++;
-            if (gCurrentPinballGame->stageTimer == 50)
-            {
-                gCurrentPinballGame->boardSubState = 0;
-                RequestBoardStateTransition(1);
-            }
-        }
-
-        if (gCurrentPinballGame->stageTimer < 29)
-        {
-            if (gCurrentPinballGame->stageTimer < 8)
-                var0 = gCurrentPinballGame->stageTimer / 4;
-            else
-                var0 = ((gCurrentPinballGame->stageTimer - 8) / 3) + 2;
-
-            group = gMain.fieldSpriteGroups[34];
-            if (group->available)
-            {
-                group->baseX = 96u - gCurrentPinballGame->cameraXOffset;
-                group->baseY = gCurrentPinballGame->rouletteSubOffset + 300u - gCurrentPinballGame->cameraYOffset;
-                if (group->baseY >= 200)
-                    group->baseY = 200;
-
-                for (i = 0; i < 9; i++)
-                {
-                    oamSimple = &group->oam[i];
-                    dst = (u16 *)&gOamBuffer[oamSimple->oamId];
-                    *dst++ = gAreaRouletteOamFramesets[var0][i * 3 + 0];
-                    *dst++ = gAreaRouletteOamFramesets[var0][i * 3 + 1];
-                    *dst++ = gAreaRouletteOamFramesets[var0][i * 3 + 2];
-
-                    gOamBuffer[oamSimple->oamId].x += group->baseX;
-                    gOamBuffer[oamSimple->oamId].y += group->baseY;
-                }
-            }
-
-            if (gCurrentPinballGame->stageTimer == 0x1C) {
-                gMain.fieldSpriteGroups[34]->available = 0;
-                gCurrentPinballGame->activePortraitType = 0;
-            }
-        }
-
-        gCurrentPinballGame->portraitDisplayState = 0;
-        break;
-    }
-}
-
-void UpdateRouletteAnimState(void)
-{
-    s16 index;
-
-    index = (gMain.systemFrameCount % 100) / 10;
-    gCurrentPinballGame->hudAnimFrameCounter = 0;
-    gCurrentPinballGame->catchArrowProgress = gBumperAnimFrames[index];
-    gCurrentPinballGame->evoArrowProgress = gBumperAnimFrames[index];
-    gCurrentPinballGame->coinRewardLevel = gBumperAnimFrames[index];
-    if (index == 7 || index == 9)
-    {
-        gCurrentPinballGame->catchArrowPaletteActive = 1;
-        gCurrentPinballGame->rouletteSlotActive = 1;
-        gCurrentPinballGame->evoArrowPaletteActive = 1;
-    }
-    else
-    {
-        gCurrentPinballGame->catchArrowPaletteActive = 0;
-        gCurrentPinballGame->rouletteSlotActive = 0;
-        gCurrentPinballGame->evoArrowPaletteActive = 0;
-    }
-}
-
-void SetRouletteActiveState(s16 arg0)
-{
-    if (arg0)
-    {
-        gCurrentPinballGame->rouletteSlotActive = 1;
-        gCurrentPinballGame->catchArrowPaletteActive = 1;
-        gCurrentPinballGame->catchProgressFlashing = 1;
-        gCurrentPinballGame->evoArrowProgress = 3;
-        gCurrentPinballGame->coinRewardLevel = 3;
-        gCurrentPinballGame->catchArrowProgress = 3;
-        gCurrentPinballGame->evoArrowPaletteActive = 1;
-    }
-    else
-    {
-        gCurrentPinballGame->evoArrowPaletteActive = 0;
-        gCurrentPinballGame->rouletteSlotActive = 0;
-        gCurrentPinballGame->catchArrowPaletteActive = 0;
-        gCurrentPinballGame->evoArrowProgress = 0;
-        gCurrentPinballGame->coinRewardLevel = 0;
-        gCurrentPinballGame->catchArrowProgress = 2;
-    }
-
-    gCurrentPinballGame->progressLevel = 1;
-}
-
-void CleanupTravelModeState(void)
-{
-    gCurrentPinballGame->travelRouletteSlotHitType = 0;
-    gCurrentPinballGame->seedotCount = 0;
-    if (gMain.selectedField == FIELD_RUBY)
-    {
-        gCurrentPinballGame->gulpinCurrentLevel = 0;
-        gCurrentPinballGame->gulpinAnimFrameIndex = 0;
-        gCurrentPinballGame->gulpinAnimFrameTimer = 0;
-    }
-
-    gCurrentPinballGame->seedotExitSequenceActive = 1;
-    gCurrentPinballGame->seedotExitSequenceTimer = 0;
-    LoadPortraitGraphics(0, 0);
-    gCurrentPinballGame->portraitDisplayState = 0;
-    gMain.fieldSpriteGroups[13]->available = 0;
-    gCurrentPinballGame->trapAnimState = 0;
-    gCurrentPinballGame->bonusTrapEnabled = 0;
-    gCurrentPinballGame->prevTravelArrowTiles[0] = gCurrentPinballGame->travelArrowTiles[0] = 0;
-    gCurrentPinballGame->prevTravelArrowTiles[1] = gCurrentPinballGame->travelArrowTiles[1] = 0;
-    gCurrentPinballGame->prevTravelArrowTiles[2] = gCurrentPinballGame->travelArrowTiles[2] = 0;
-    ResetEventState();
-}
-
-void InitTravelMode(void)
-{
-    gCurrentPinballGame->boardSubState = 0;
-    gCurrentPinballGame->stageTimer = 0;
-    gCurrentPinballGame->boardModeType = 2;
-    gCurrentPinballGame->eventTimer = gCurrentPinballGame->timerBonus + 3600;
-    gCurrentPinballGame->timerBonus = 0;
-    gCurrentPinballGame->saverTimeRemaining = 1800;
-    DmaCopy16(3, gDefaultBallPalette, (void *)0x05000180, 0x20);
-}
-
-void UpdateTravelMode(void)
-{
-    s16 var0;
-
-    if (gCurrentPinballGame->boardModeType && gCurrentPinballGame->eventTimer < 2 && gCurrentPinballGame->boardSubState < 6)
-    {
-        m4aMPlayAllStop();
-        m4aSongNumStart(MUS_END_OF_BALL2);
-        gCurrentPinballGame->stageTimer = 200;
-        gCurrentPinballGame->boardSubState = 6;
-    }
-
-    switch (gCurrentPinballGame->boardSubState)
-    {
-    case 0:
-        if (gMain.modeChangeFlags == MODE_CHANGE_NONE)
-        {
-            gCurrentPinballGame->boardSubState++;
-            gCurrentPinballGame->portraitCycleFrame = 0;
-        }
-        gCurrentPinballGame->travelRouletteSlotHitType = 0;
-        break;
-    case 1:
-        LoadPortraitGraphics(4, 0);
-        if (gCurrentPinballGame->stageTimer == 35)
-            m4aSongNumStart(MUS_TRAVEL_MODE);
-
-        if (gCurrentPinballGame->travelRouletteSlotHitType)
-        {
-            gCurrentPinballGame->boardSubState++;
-            if (gCurrentPinballGame->stageTimer < 35)
-                m4aSongNumStart(MUS_TRAVEL_MODE);
-
-            gCurrentPinballGame->seedotCount = 0;
-            gCurrentPinballGame->seedotExitSequenceActive = 1;
-            gCurrentPinballGame->seedotExitSequenceTimer = 0;
-        }
-
-        var0 = (gCurrentPinballGame->stageTimer % 40) / 10;
-        gCurrentPinballGame->stageTimer++;
-        if (var0 == 0)
-        {
-            gCurrentPinballGame->travelArrowTiles[0] = 0;
-            gCurrentPinballGame->travelArrowTiles[1] = 0;
-            gCurrentPinballGame->travelArrowTiles[2] = 0;
-        }
-        else if (var0 == 1)
-        {
-            gCurrentPinballGame->travelArrowTiles[0] = 3;
-            gCurrentPinballGame->travelArrowTiles[1] = 1;
-            gCurrentPinballGame->travelArrowTiles[2] = 0;
-        }
-        else if (var0 == 2)
-        {
-            gCurrentPinballGame->travelArrowTiles[0] = 3;
-            gCurrentPinballGame->travelArrowTiles[1] = 3;
-            gCurrentPinballGame->travelArrowTiles[2] = 2;
-        }
-        else
-        {
-            gCurrentPinballGame->travelArrowTiles[0] = 3;
-            gCurrentPinballGame->travelArrowTiles[1] = 3;
-            gCurrentPinballGame->travelArrowTiles[2] = 3;
-        }
-
-        gCurrentPinballGame->prevTravelArrowTiles[0] = gCurrentPinballGame->travelArrowTiles[0];
-        gCurrentPinballGame->prevTravelArrowTiles[1] = gCurrentPinballGame->travelArrowTiles[1];
-        gCurrentPinballGame->prevTravelArrowTiles[2] = gCurrentPinballGame->travelArrowTiles[2];
-        break;
-    case 2:
-        ShowBonusTrapSprite();
-        gCurrentPinballGame->trapAnimState = 2;
-        LoadPortraitGraphics(0, 0);
-        gCurrentPinballGame->prevTravelArrowTiles[0] = gCurrentPinballGame->travelArrowTiles[0] = 0;
-        gCurrentPinballGame->prevTravelArrowTiles[1] = gCurrentPinballGame->travelArrowTiles[1] = 0;
-        gCurrentPinballGame->prevTravelArrowTiles[2] = gCurrentPinballGame->travelArrowTiles[2] = 0;
-        gCurrentPinballGame->boardSubState++;
-        break;
-    case 3:
-        AnimateBonusTrapSprite();
-        if (gCurrentPinballGame->ballCatchState == 4)
-            gCurrentPinballGame->boardSubState++;
-        break;
-    case 4:
-        gCurrentPinballGame->boardModeType = 3;
-        gCurrentPinballGame->boardSubState++;
-        gCurrentPinballGame->stageTimer = 0;
-        gCurrentPinballGame->portraitCycleFrame = 0;
-        gCurrentPinballGame->modeOutcomeValues[0] = 47;
-        LoadPortraitGraphics(7, 0);
-        break;
-    case 5:
-        if (gCurrentPinballGame->modeAnimTimer == 145)
-        {
-            gCurrentPinballGame->modeAnimTimer++;
-            gCurrentPinballGame->modeOutcomeValues[0] = 47;
-            LoadPortraitGraphics(7, 0);
-            if (JOY_NEW(A_BUTTON))
-            {
-                gCurrentPinballGame->modeAnimTimer = 144;
-                m4aMPlayAllStop();
-                LoadPortraitGraphics(0, 0);
-                if (gCurrentPinballGame->areaVisitCount < 5)
-                {
-                    var0 = gCurrentPinballGame->areaRouletteFarSlot;
-                    if (gCurrentPinballGame->travelRouletteSlotHitType == 1)
-                        gCurrentPinballGame->areaRouletteSlotIndex = gCurrentPinballGame->areaRouletteNextSlot;
-                    else
-                        gCurrentPinballGame->areaRouletteSlotIndex = gCurrentPinballGame->areaRouletteFarSlot;
-
-                    gCurrentPinballGame->areaRouletteNextSlot = (var0 + 1) % 6;
-                    gCurrentPinballGame->areaRouletteFarSlot = (var0 + 2) % 6;
-                    gCurrentPinballGame->areaVisitCount++;
-                }
-                else
-                {
-                    gCurrentPinballGame->areaRouletteSlotIndex = 6;
-                    gCurrentPinballGame->areaVisitCount = 0;
-                }
-            }
-            else if (JOY_NEW(B_BUTTON))
-            {
-                m4aMPlayAllStop();
-                m4aSongNumStart(SE_MENU_CANCEL);
-                gCurrentPinballGame->modeAnimTimer = 60;
-                gCurrentPinballGame->boardSubState = 6;
-                if (gCurrentPinballGame->allHolesLit)
-                    gCurrentPinballGame->allHolesLitDelayTimer = 120;
-            }
-        }
-
-        if (gCurrentPinballGame->modeAnimTimer == 130)
-        {
-            gCurrentPinballGame->modeAnimTimer++;
-            if (gCurrentPinballGame->stageTimer < 490)
-            {
-                RunTravelEventCutscene();
-                if (gCurrentPinballGame->scoreCounterAnimationEnabled && gCurrentPinballGame->stageTimer >= 428)
-                    gCurrentPinballGame->stageTimer = 428;
-
-                if (gCurrentPinballGame->stageTimer == 394)
-                {
-                    gCurrentPinballGame->scoreCounterAnimationEnabled = TRUE;
-                    gCurrentPinballGame->scoreAddedInFrame = 500000;
-                }
-            }
-            else
-            {
-                gCurrentPinballGame->modeAnimTimer = 120;
-            }
-
-            gCurrentPinballGame->stageTimer++;
-        }
-        else if (gCurrentPinballGame->modeAnimTimer == 0)
-        {
-            gCurrentPinballGame->boardSubState++;
-            gCurrentPinballGame->stageTimer = 0;
-            if (gCurrentPinballGame->travelModeCompletionCount < 99)
-                gCurrentPinballGame->travelModeCompletionCount++;
-        }
-        break;
-    case 6:
-        AnimateBonusTrapSprite();
-        gMain.fieldSpriteGroups[13]->available = 0;
-        CleanupTravelModeState();
-        gCurrentPinballGame->boardSubState++;
-        break;
-    case 7:
-        if (gCurrentPinballGame->stageTimer)
-        {
-            gCurrentPinballGame->stageTimer--;
-        }
-        else
-        {
-            RequestBoardStateTransition(1);
-            gCurrentPinballGame->boardSubState = 0;
-        }
-        break;
-    }
-}
-
-//Evolution_mode.c starts here
-void CleanupEvolutionModeState(void)
-{
-    s16 i;
-
-    LoadPortraitGraphics(0, 0);
-    gCurrentPinballGame->portraitDisplayState = 0;
-    for (i = 0; i < 3; i++)
-    {
-        if (i < gCurrentPinballGame->evoItemCount)
-            gCurrentPinballGame->catchLights[i] = 1;
-        else
-            gCurrentPinballGame->catchLights[i] = 0;
-    }
-
-    gMain.fieldSpriteGroups[13]->available = 0;
-    gCurrentPinballGame->trapAnimState = 0;
-    gCurrentPinballGame->bonusTrapEnabled = 0;
-    ResetEventState();
-}
-
-void InitEvolutionMode(void)
-{
-    gCurrentPinballGame->boardSubState = 0;
-    gCurrentPinballGame->stageTimer = 0;
-    gCurrentPinballGame->boardModeType = 2;
-    gCurrentPinballGame->eventTimer = gCurrentPinballGame->timerBonus + 7200;
-    gCurrentPinballGame->timerBonus = 0;
-    gCurrentPinballGame->saverTimeRemaining = 3600;
-    if (gCurrentPinballGame->currentSpecies == SPECIES_WURMPLE)
-    {
-        gCurrentPinballGame->evoItemGfxIndex = 0;
-    }
-    else if (gCurrentPinballGame->currentSpecies == SPECIES_GLOOM)
-    {
-        if (gMain.selectedField == FIELD_RUBY)
-            gCurrentPinballGame->evoItemGfxIndex = 1;
-        else
-            gCurrentPinballGame->evoItemGfxIndex = 7;
-    }
-    else if (gCurrentPinballGame->currentSpecies == SPECIES_CLAMPERL)
-    {
-        gCurrentPinballGame->evoItemGfxIndex = 3;
-    }
-    else
-    {
-        gCurrentPinballGame->evoItemGfxIndex = gSpeciesInfo[gCurrentPinballGame->currentSpecies].evolutionMethod - 1;
-        if (gCurrentPinballGame->evoItemGfxIndex < 0)
-            gCurrentPinballGame->evoItemGfxIndex = 0;
-    }
-
-    DmaCopy16(3, gDefaultBallPalette, (void *)0x05000180, 0x20);
-    gCurrentPinballGame->evoArrowProgress = 0;
-    gCurrentPinballGame->rouletteSlotActive = 0;
-    gCurrentPinballGame->catchModeEventTimer = 0;
-    gCurrentPinballGame->evoItemSlotIndex = 0;
-    gCurrentPinballGame->evoItemsCaught = 0;
-    gCurrentPinballGame->evoModeResetFlag = 0;
-    gCurrentPinballGame->evoItemAppearTimer = 0;
-    LoadPortraitGraphics(3, 0);
-}
-
-void UpdateEvolutionMode(void)
-{
-    s16 i, j;
-    s16 var0;
-    struct SpriteGroup *group;
-    struct OamDataSimple *oamSimple;
-    s16 index;
-
-    if (gCurrentPinballGame->boardModeType && gCurrentPinballGame->eventTimer < 2 && gCurrentPinballGame->boardSubState < 8)
-    {
-        m4aMPlayAllStop();
-        m4aSongNumStart(MUS_END_OF_BALL2);
-        gCurrentPinballGame->stageTimer = 200;
-        gCurrentPinballGame->boardSubState = 8;
-    }
-
-    switch (gCurrentPinballGame->boardSubState)
-    {
-    case 0:
-        if (gCurrentPinballGame->evoModeShuffleRound < 2)
-        {
-            for (i = 0; i < 8; i++)
-                gCurrentPinballGame->evoItemShufflePool[i] = i;
-
-            index = (Random() + gMain.systemFrameCount) % 5;
-            gCurrentPinballGame->evoShuffledSlots[0] = gCurrentPinballGame->evoItemShufflePool[index];
-            for (i = index; i < 7; i++)
-                gCurrentPinballGame->evoItemShufflePool[i] = gCurrentPinballGame->evoItemShufflePool[i + 1];
-
-            index = (Random() + gMain.systemFrameCount) % 4;
-            gCurrentPinballGame->evoShuffledSlots[1] = gCurrentPinballGame->evoItemShufflePool[index];
-            for (i = index; i < 6; i++)
-                gCurrentPinballGame->evoItemShufflePool[i] = gCurrentPinballGame->evoItemShufflePool[i + 1];
-
-            index = (Random() + gMain.systemFrameCount) % 3;
-            gCurrentPinballGame->evoShuffledSlots[2] = gCurrentPinballGame->evoItemShufflePool[index];
-        }
-        else
-        {
-            for (i = 0; i < 8; i++)
-                gCurrentPinballGame->evoItemShufflePool[i] = i;
-
-            index = (Random() + gMain.systemFrameCount) % 6;
-            gCurrentPinballGame->evoShuffledSlots[0] = gCurrentPinballGame->evoItemShufflePool[index];
-            for (i = index; i < 7; i++)
-                gCurrentPinballGame->evoItemShufflePool[i] = gCurrentPinballGame->evoItemShufflePool[i + 1];
-
-            index = (Random() + gMain.systemFrameCount) % 7;
-            gCurrentPinballGame->evoShuffledSlots[1] = gCurrentPinballGame->evoItemShufflePool[index];
-            for (i = index; i < 6; i++)
-                gCurrentPinballGame->evoItemShufflePool[i] = gCurrentPinballGame->evoItemShufflePool[i + 1];
-
-            index = (Random() + gMain.systemFrameCount) % 6;
-            gCurrentPinballGame->evoShuffledSlots[2] = gCurrentPinballGame->evoItemShufflePool[index];
-            for (i = index; i < 5; i++)
-                gCurrentPinballGame->evoItemShufflePool[i] = gCurrentPinballGame->evoItemShufflePool[i + 1];
-
-            if (gMain.selectedField == FIELD_SAPPHIRE && gCurrentPinballGame->numCompletedBonusStages < 5)
-            {
-                for (i = 0; i < 3; i++)
-                {
-                    if (gCurrentPinballGame->evoShuffledSlots[i] == 5)
-                    {
-                        index = (Random() + gMain.systemFrameCount) % 5;
-                        gCurrentPinballGame->evoShuffledSlots[i] = gCurrentPinballGame->evoItemShufflePool[index];
-                    }
-                }
-            }
-        }
-
-        gCurrentPinballGame->evoModeShuffleRound++;
-        gCurrentPinballGame->boardSubState++;
-        break;
-    case 1:
-        gCurrentPinballGame->evoItemSlotIndex = gCurrentPinballGame->evoShuffledSlots[gCurrentPinballGame->evoItemsCaught];
-        gCurrentPinballGame->evoItemPosX = gEvoItemPositions[gMain.selectedField][gCurrentPinballGame->evoItemSlotIndex].x;
-        gCurrentPinballGame->evoItemPosY = gEvoItemPositions[gMain.selectedField][gCurrentPinballGame->evoItemSlotIndex].y;
-        gCurrentPinballGame->evoItemAppearTimer = 80;
-        gMain.fieldSpriteGroups[40]->available = 1;
-        gCurrentPinballGame->boardSubState++;
-        break;
-    case 2:
-        UpdateEvolutionItemAnimation();
-        gCurrentPinballGame->stageTimer = 0;
-        break;
-    case 3:
-        gCurrentPinballGame->trapAnimState = 1;
-        if (gCurrentPinballGame->stageTimer < 8)
-        {
-            gCurrentPinballGame->stageTimer++;
-        }
-        else
-        {
-            ShowBonusTrapSprite();
-            gCurrentPinballGame->trapAnimState = 2;
-            gCurrentPinballGame->boardSubState++;
-        }
-        break;
-    case 4:
-        AnimateBonusTrapSprite();
-        if (gCurrentPinballGame->ballCatchState == 4)
-            gCurrentPinballGame->boardSubState++;
-        break;
-    case 5:
-        gCurrentPinballGame->boardModeType = 3;
-        gCurrentPinballGame->preEvoSpecies = gCurrentPinballGame->currentSpecies;
-        RegisterCaptureOrEvolution(1);
-        gCurrentPinballGame->postEvoSpecies = gCurrentPinballGame->currentSpecies;
-        gCurrentPinballGame->stageTimer = 0;
-        gCurrentPinballGame->boardSubState++;
-        break;
-    case 6:
-        if (gCurrentPinballGame->modeAnimTimer == 148)
-        {
-            gCurrentPinballGame->modeAnimTimer++;
-            if (gMain.spriteGroups[13].available)
-            {
-                if (gCurrentPinballGame->chikoritaProjectileTimer >= 80)
-                {
-                    RunEvolutionCutscene();
-                    if (gCurrentPinballGame->boardSubState == 6)
-                        gCurrentPinballGame->stageTimer++;
-                }
-            }
-            else
-            {
-                RunEvolutionCutscene();
-                if (gCurrentPinballGame->boardSubState == 6)
-                    gCurrentPinballGame->stageTimer++;
-            }
-        }
-        break;
-    case 7:
-        if (gCurrentPinballGame->modeAnimTimer == 148)
-        {
-            gCurrentPinballGame->modeAnimTimer++;
-            if (gCurrentPinballGame->stageTimer < 280)
-            {
-                if (gCurrentPinballGame->stageTimer == 0)
-                {
-                    InitEvolutionSuccessDisplay();
-                    m4aSongNumStart(MUS_SUCCESS2);
-                }
-                else
-                {
-                    AnimateEvolutionSuccessScreen();
-                    if (gCurrentPinballGame->stageTimer < 230)
-                    {
-                        if (gCurrentPinballGame->nameRevealAnimFrame == 150)
-                            gCurrentPinballGame->nameRevealAnimFrame--;
-                    }
-
-                    if (gCurrentPinballGame->stageTimer == 270)
-                    {
-                        LoadPortraitGraphics(0, 0);
-                        gCurrentPinballGame->portraitDisplayState = 0;
-                    }
-
-                    if (gCurrentPinballGame->scoreCounterAnimationEnabled)
-                        gCurrentPinballGame->stageTimer = 181;
-
-                    if (gCurrentPinballGame->stageTimer == 180) {
-                        gCurrentPinballGame->scoreCounterAnimationEnabled = TRUE;
-                        gCurrentPinballGame->scoreAddedInFrame = 5000000;
-                    }
-                }
-
-                if (gCurrentPinballGame->stageTimer < 30)
-                {
-                    gMain.scoreOverlayActive = 1;
-                    var0 = gCurrentPinballGame->stageTimer;
-                    gCurrentPinballGame->cutsceneTilemapColumn = gCurrentPinballGame->stageTimer;
-                    for (i = 0; i <= var0; i++)
-                    {
-                        for (j = 2; j < 12; j++)
-                            gBG0TilemapBuffer[((j + 15) * 0x20) + i] = 0xC100;
-                    }
-
-                    DmaCopy16(3, gBG0TilemapBuffer, (void *)0x06002000, 0x800);
-                }
-
-                if (gCurrentPinballGame->stageTimer >= 240 && gCurrentPinballGame->stageTimer < 270)
-                {
-                    var0 = gCurrentPinballGame->stageTimer - 240;
-                    for (i = 0; i <= var0; i ++)
-                    {
-                        for (j = 2; j < 12; j++)
-                            gBG0TilemapBuffer[((j + 15) << 5) + i] = 0x1FF;
-                    }
-
-                    DmaCopy16(3, gBG0TilemapBuffer, (void *)0x06002000, 0x800);
-                    if (gCurrentPinballGame->stageTimer == 269)
-                    {
-                        gMain.scoreOverlayActive = 0;
-                        gMain.blendControl = 0;
-                        gMain.blendBrightness = 0;
-                        gMain.blendAlpha = 0;
-                    }
-                }
-
-                gCurrentPinballGame->stageTimer++;
-            }
-            else
-            {
-                gCurrentPinballGame->modeAnimTimer = 40;
-            }
-        }
-        else if (gCurrentPinballGame->modeAnimTimer == 24)
-        {
-            if (gCurrentPinballGame->evoItemCount < 3)
-            {
-                gCurrentPinballGame->evoCatchLightSlot1 = gCurrentPinballGame->evoItemCount;
-                gCurrentPinballGame->evoCatchLightSlot2 = gCurrentPinballGame->evoItemCount;
-                gCurrentPinballGame->catchLights[gCurrentPinballGame->evoCatchLightSlot1] = 1;
-                gCurrentPinballGame->evoBlinkTimer = 120;
-                gCurrentPinballGame->evoItemCount++;
-                if (gCurrentPinballGame->evoItemCount < 3)
-                {
-                    gCurrentPinballGame->evoCatchLightSlot2 = gCurrentPinballGame->evoItemCount;
-                    gCurrentPinballGame->catchLights[gCurrentPinballGame->evoCatchLightSlot2] = 1;
-                    gCurrentPinballGame->evoBlinkTimer = 120;
-                    gCurrentPinballGame->evoItemCount++;
-                }
-            }
-
-            if (gCurrentPinballGame->currentSpecies == SPECIES_NINJASK)
-            {
-                gCurrentPinballGame->caughtMonCount++;
-                if (gCurrentPinballGame->bonusMonCatchCount < 99)
-                    gCurrentPinballGame->bonusMonCatchCount++;
-
-                if (gCurrentPinballGame->caughtMonCount == 15)
-                    gCurrentPinballGame->oneUpAnimTimer = 92;
-            }
-
-            gCurrentPinballGame->caughtMonCount++;
-            if (gCurrentPinballGame->caughtMonCount == 15)
-                gCurrentPinballGame->oneUpAnimTimer = 92;
-
-            if (gCurrentPinballGame->bonusMonCatchCount < 99)
-                gCurrentPinballGame->bonusMonCatchCount++;
-
-            gCurrentPinballGame->boardSubState++;
-            gCurrentPinballGame->stageTimer = 0;
-        }
-        break;
-    case 8:
-        group = gMain.fieldSpriteGroups[32];
-        oamSimple = &group->oam[0];
-        gOamBuffer[oamSimple->oamId].x = oamSimple->xOffset;
-        gOamBuffer[oamSimple->oamId].y = oamSimple->yOffset - 56;
-
-        gMain.fieldSpriteGroups[32]->available = 0;
-        UpdateEvolutionItemAnimation();
-        gMain.fieldSpriteGroups[40]->available = 0;
-        gCurrentPinballGame->activePortraitType = 0;
-        AnimateBonusTrapSprite();
-        gMain.fieldSpriteGroups[13]->available = 0;
-        gCurrentPinballGame->shopTransitionActive = 1;
-        gCurrentPinballGame->shopAnimTimer = 0;
-        gCurrentPinballGame->evolutionShopActive = 0;
-        gCurrentPinballGame->boardSubState++;
-        if (gCurrentPinballGame->allHolesLit)
-            gCurrentPinballGame->allHolesLitDelayTimer = 120;
-        break;
-    case 9:
-        CleanupEvolutionModeState();
-        gCurrentPinballGame->boardSubState++;
-        break;
-    case 10:
-        if (gCurrentPinballGame->stageTimer)
-        {
-            gCurrentPinballGame->stageTimer--;
-        }
-        else
-        {
-            if (gCurrentPinballGame->catchLights[2] == 1)
-                RequestBoardStateTransition(3);
-            else
-                RequestBoardStateTransition(1);
-
-            gCurrentPinballGame->boardSubState = 0;
-        }
-        break;
-    }
-}
-
-void UpdateEvolutionItemAnimation(void)
-{
-    s16 i;
-    struct SpriteGroup *group;
-    struct OamDataSimple *oamSimple;
-    u16 *dst;
-    const u16 *src;
-    struct Vector32 tempVector;
-    int xx, yy;
-    int squaredMagnitude;
-    s16 index;
-
-    index = (gMain.systemFrameCount % 75) / 3;
-    group = gMain.fieldSpriteGroups[32];
-    if (gCurrentPinballGame->evoItemAppearTimer)
-    {
-        group = gMain.fieldSpriteGroups[40];
-        if (gCurrentPinballGame->evoItemAppearTimer == 80)
-        {
-            gCurrentPinballGame->activePortraitType = 15;
-            DmaCopy16(3, gEvoItemAppear_GfxList[gCurrentPinballGame->evoItemGfxIndex], (void *)0x06015800, 0x1C00);
-            DmaCopy16(3, &gEvoItemPalettes[gCurrentPinballGame->evoItemGfxIndex], (void *)0x050003E0, 0x20);
-            gCurrentPinballGame->evoItemAnimFrame = 0;
-            gCurrentPinballGame->evoItemAnimFrameTimer = 0;
-        }
-
-        if (gCurrentPinballGame->evoItemAppearTimer == 60)
-            MPlayStart(&gMPlayInfo_SE1, &se_evo_item_appear);
-
-        if (gEvoItemAppearFrameThresholds[gCurrentPinballGame->evoItemAnimFrame] > gCurrentPinballGame->evoItemAnimFrameTimer)
-        {
-            gCurrentPinballGame->evoItemAnimFrameTimer++;
-        }
-        else
-        {
-            gCurrentPinballGame->evoItemAnimFrameTimer = 1;
-            gCurrentPinballGame->evoItemAnimFrame++;
-            if (gCurrentPinballGame->evoItemAnimFrame == 29)
-            {
-                gCurrentPinballGame->evoItemAnimFrame = 28;
-                gCurrentPinballGame->evoItemAppearTimer = 1;
-                gMain.fieldSpriteGroups[40]->available = 0;
-                gMain.fieldSpriteGroups[32]->available = 1;
-                MPlayStart(&gMPlayInfo_SE1, &se_evo_item_finish_appear);
-                gCurrentPinballGame->activePortraitType = 0;
-            }
-        }
-
-        group->baseX = gCurrentPinballGame->evoItemPosX - gCurrentPinballGame->cameraXOffset + 8;
-        group->baseY = gCurrentPinballGame->evoItemPosY - gCurrentPinballGame->cameraYOffset + 8;
-        if (group->baseY < -30)
-            group->baseY = -30;
-        if (group->baseY > 200)
-            group->baseY = 200;
-
-        for (i = 0; i < 5; i++)
-        {
-            oamSimple = &group->oam[i];
-            dst = (u16 *)&gOamBuffer[oamSimple->oamId];
-            *dst++ = gEvoItemAnimOamFramesets[gCurrentPinballGame->evoItemAnimFrame][i * 3 + 0];
-            *dst++ = gEvoItemAnimOamFramesets[gCurrentPinballGame->evoItemAnimFrame][i * 3 + 1];
-            *dst++ = gEvoItemAnimOamFramesets[gCurrentPinballGame->evoItemAnimFrame][i * 3 + 2];
-
-            gOamBuffer[oamSimple->oamId].x += group->baseX;
-            gOamBuffer[oamSimple->oamId].y += group->baseY;
-        }
-
-        gCurrentPinballGame->evoItemAppearTimer--;
-    }
-    else
-    {
-        tempVector.x = gCurrentPinballGame->ball->positionQ0.x - (gCurrentPinballGame->evoItemPosX + 8);
-        tempVector.y = gCurrentPinballGame->ball->positionQ0.y - (gCurrentPinballGame->evoItemPosY + 8);
-        xx = tempVector.x * tempVector.x;
-        yy = tempVector.y * tempVector.y;
-        squaredMagnitude = xx + yy;
-        if (squaredMagnitude < 82 &&
-            ((gCurrentPinballGame->boardLayerDepth == 0 && gCurrentPinballGame->evoItemSlotIndex <= 5) || (gCurrentPinballGame->boardLayerDepth == 2 && gCurrentPinballGame->evoItemSlotIndex > 5)) &&
-            gCurrentPinballGame->evoItemsCaught < 3)
-        {
-            gCurrentPinballGame->scoreAddedInFrame = 10000;
-            MPlayStart(&gMPlayInfo_SE1, &se_evo_item_collected);
-            gCurrentPinballGame->boardSubState = 1;
-            gCurrentPinballGame->catchLights[gCurrentPinballGame->evoItemsCaught] = 5;
-            gCurrentPinballGame->evoItemsCaught++;
-            gMain.fieldSpriteGroups[32]->available = 0;
-            if (gCurrentPinballGame->evoItemsCaught == 3)
-            {
-                gCurrentPinballGame->evoItemsCaught = 0;
-                gCurrentPinballGame->boardSubState = 3;
-            }
-        }
-
-        group->baseX = gCurrentPinballGame->evoItemPosX - gCurrentPinballGame->cameraXOffset - 8;
-        group->baseY = gCurrentPinballGame->evoItemPosY - gCurrentPinballGame->cameraYOffset - 8;
-        if (group->baseY < -30)
-            group->baseY = -30;
-        if (group->baseY > 200)
-            group->baseY = 200;
-
-        if (index > 14)
-            index = 14;
-
-        DmaCopy16(3, gEvoItemTilesGfxPtrs[gCurrentPinballGame->evoItemGfxIndex] + index * 0x200, (void *)0x060116C0, 0x200);
-        oamSimple = &group->oam[0];
-        gOamBuffer[oamSimple->oamId].x = oamSimple->xOffset + group->baseX;
-        gOamBuffer[oamSimple->oamId].y = oamSimple->yOffset + group->baseY;
-    }
-}
-
-//catch_mode.c starts here
-void CleanupCatchEmState(void)
-{
-    s16 i;
-
-    gCurrentPinballGame->creatureHitCount = 0;
-    gCurrentPinballGame->captureFlashTimer = 0;
-    gMain.fieldSpriteGroups[18]->available = 0;
-    gMain.fieldSpriteGroups[12]->available = 0;
-    gCurrentPinballGame->jirachiCollisionEnabled = 0;
-    LoadPortraitGraphics(0, 0);
-    gCurrentPinballGame->portraitDisplayState = 0;
-    ResetEventState();
-    for (i = 0; i < 6; i++)
-        gCurrentPinballGame->hatchTilePalette[i] = 13;
-
-    for (i = 0; i < 3; i++)
-    {
-        if (i < gCurrentPinballGame->evoItemCount)
-            gCurrentPinballGame->catchLights[i] = 1;
-        else
-            gCurrentPinballGame->catchLights[i] = 0;
-    }
-}
-
-void InitCatchEmMode(void)
-{
-    s16 i, j;
-
-    gCurrentPinballGame->boardSubState = 0;
-    gCurrentPinballGame->stageTimer = 0;
-    gCurrentPinballGame->boardModeType = 1;
-    gCurrentPinballGame->eventTimer = gCurrentPinballGame->timerBonus + BONUS_CATCH_TIME;
-    gCurrentPinballGame->timerBonus = 0;
-    gCurrentPinballGame->creatureHitCount = 0;
-    gCurrentPinballGame->creatureHitCooldown = 0;
-    gCurrentPinballGame->captureFlashTimer = 0;
-    gCurrentPinballGame->hatchTilesBoardAcknowledged = 0;
-    gCurrentPinballGame->hatchSequentialTilesRevealed = 0;
-    gCurrentPinballGame->hatchTilesBumperAcknowledged = 0;
-    gCurrentPinballGame->hatchSequentialTileRevealFrameAnimTimer = 0;
-    gCurrentPinballGame->hatchFrameId = 0;
-    gCurrentPinballGame->catchArrowProgress = 0;
-    gCurrentPinballGame->catchProgressFlashing = 0;
-
-    if (gCurrentPinballGame->catchEmModeStartCount == 0)
-    {
-        gCurrentPinballGame->saverTimeRemaining = 6000;
-    }
-    else
-    {
-        gCurrentPinballGame->saverTimeRemaining = 4200;
-    }
-    gCurrentPinballGame->catchEmModeStartCount++;
-
-    DmaCopy16(3, gDefaultBallPalette, (void *)PLTT + 0x180, 0x20);
-
-    for (i = 0; i < 6; i++)
-    {
-        gCurrentPinballGame->hatchTileShufflePool[i] = i;
-    }
-
-    gCurrentPinballGame->hatchGridCellIndex = gMain.systemFrameCount % 6;
-    gCurrentPinballGame->hatchTilesRemaining = 5;
-
-    for (j = gCurrentPinballGame->hatchGridCellIndex; j < gCurrentPinballGame->hatchTilesRemaining; j++)
-    {
-        gCurrentPinballGame->hatchTileShufflePool[j] = gCurrentPinballGame->hatchTileShufflePool[j+1];
     }
 }
